% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddrum.R
\docType{data}
\name{reddrum}
\alias{reddrum}
\title{RAD-seq genotypes for 400 red drum from the U.S. Gulf of Mexico and Atlantic}
\format{
An object of class \code{list} of length 2.
}
\source{

}
\usage{
reddrum
}
\description{
A genind object with 400 individuals genotypes at 1,539 loci from 11 sampled localities
}
\references{

}
\keyword{datasets}
