% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twentygen.R
\docType{data}
\name{twentygen}
\alias{twentygen}
\title{Twenty generation of simulated genotype data}
\format{
An object of class \code{genind} of length 1.
}
\source{

}
\usage{
twentygen
}
\description{
genind object with 100 loci from a population (N=10) simulated for 20 generations
}
\references{

}
\keyword{datasets}
